move to release_util
