+++ pyartcd/pyartcd/pipelines/prepare_release.py
@@ -18,7 +18,7 @@ from tenacity import retry, stop_after_attempt, wait_fixed
 from datetime import datetime, timedelta
 
 from artcommonlib.assembly import AssemblyTypes, assembly_group_config
-from artcommonlib.util import get_assembly_release_date
+from artcommonlib.release_util import get_assembly_release_date
 from elliottlib.errata import set_blocking_advisory, get_blocking_advisories
 from elliottlib.errata import get_brew_builds
 from elliottlib.model import Model
